requirements.txt
asgiref==3.8.1
Django==4.2.11
django-phonenumber-field==7.3.0
phonenumbers==8.13.35
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
